Automatically append the "enable" rule to the ruleset
We expect that 9 times out of 10 the user will want to enable the audit
subsystem with an `-e 1` rule at the end of the ruleset. Add a class
parameter to the base class to control this. The enable action can be
overridden to `-e 0` (disable), `-e 2` (lock), or removed altogether,
so the module only configures rules as given explicitly via
`goaudit::rule` resources.